api/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pyyaml
pydantic-settings